store config in KV

+++ scripts/typegen.sh
@@ -1,7 +1,11 @@
 run_cf_typegen() {
     folder="$1"
     if [ -d "$folder" ]; then
-        cp .dev.vars "$folder"
+        if [ -f "$folder/.dev.vars" ]; then
+            awk 'NR==FNR{a[$1];next}!($1 in a)' .dev.vars "$folder/.dev.vars" >> "$folder/.dev.vars"
+        else
+            cp .dev.vars "$folder"
+        fi
         (cd "$folder" && npm run cf-typegen)
     else
         echo "Directory $folder does not exist."
